requirements.txt
requests
pytube
spotipy
beautifulsoup4
lxml
yt-dlp
PySide6
Pillow
streamlink